refractor: improve logic for accessibility of operation on data

+++ src/main/java/com/project/hearmeout_backend/service/CommentService.java
@@ -2,6 +2,7 @@ package com.project.hearmeout_backend.service;
 
 import com.project.hearmeout_backend.dto.request.comment_request.CommentRequestDTO;
 import com.project.hearmeout_backend.exception.CommentNotFoundException;
+import com.project.hearmeout_backend.exception.InvalidOperationException;
 import com.project.hearmeout_backend.exception.PostNotFoundException;
 import com.project.hearmeout_backend.exception.UserNotFoundException;
 import com.project.hearmeout_backend.mapper.CommentMapper;
@@ -22,9 +23,9 @@ public class CommentService {
     private final CommentRepository commentRepo;
 
     @Transactional
-    public void createNewComment(CommentRequestDTO commentRequestDTO)
+    public void createNewComment(CommentRequestDTO commentRequestDTO, Long userId)
             throws UserNotFoundException, PostNotFoundException {
-        User author = userService.checkAndGetUserByUserId(commentRequestDTO.getAuthorId());
+        User author = userService.checkAndGetUserByUserId(userId);
         Post post = postService.checkAndGetPost(commentRequestDTO.getPostId());
 
         Comment newComment = CommentMapper.toCommentEntity(commentRequestDTO, post, author);
@@ -33,9 +34,13 @@ public class CommentService {
     }
 
     @Transactional
-    public void removeComment(Long commentId)
+    public void removeComment(Long commentId, Long userId)
             throws CommentNotFoundException {
         Comment comment = checkAndGetComment(commentId);
+
+        if(!comment.getAuthor().getId().equals(userId)) {
+            throw new InvalidOperationException("Operation only allowed for account owner");
+        }
 
         commentRepo.delete(comment);
     }
@@ -49,9 +54,14 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateCommentBody(Long commentId, String body)
+    public void updateCommentBody(Long commentId, String body, Long userId)
             throws CommentNotFoundException {
         Comment comment = checkAndGetComment(commentId);
+
+        if(!comment.getAuthor().getId().equals(userId)) {
+            throw new InvalidOperationException("Operation only allowed for account owner");
+        }
+
         comment.setBody(body);
 
         commentRepo.save(comment);
